+++ syntax/tests/printer/other/ocamlString.ml
@@ -7,3 +7,21 @@ let x = "foo\x0Abar"
 let x = "foo\o012bar"
 
 let x = "😁 this works now 😆"
+let x = {|😁 this works now 😆|}
+
+
+(* The `//` should not result into an extra comment *)
+let x = {j|https://www.apple.com|j}
+let x = {|https://www.apple.com|}
+let x = {js|https://www.apple.com|js}
+let x = {|https://www.apple.com|}
+let x = {sql|https://www.apple.com|sql}
+
+(* /* */ should not result in an extra comments *)
+let x = {j|/* https://www.apple.com */|j}
+let x = {|/* https://www.apple.com*/|}
+let x = {js|/*https://www.apple.com*/|js}
+let x = {|/*https://www.apple.com*/|}
+let x = {sql|/*https://www.apple.com*/|sql}
+
+let x = {js|`https://${appleWebsite}`|js}